extracted event session merge logic
